Fix the cycle time bug
Normally, if we use usleep() to sleep for 25ms, it will not
postpone exactlly 25ms, it will will approximately 25ms.
Therefore, we use gettimeofday() in time.h to measure the exact
exection time before sending the command and use the tiny time interval
to subtract 25ms in usleep() so that we could make each cycle time to be
approach 25ms.
And also we adjust usleep() from 25000 to 24900 because for usleep()
itself will cause some time to execute the function. AKA more that 25000us
so we choose 24900 to reduce the more execution time for usleep().