Fix the real-time issue
Originally, we use usleep(25000) for make the command
sending in 25ms freqency. However, because the computation and other
function like: printf will consume some execution time and cause the
time interval over 25ms around 100us. Therefore, we use the
gettimeofday() in sys/time.h to calculate the precise execution time of
each cycle and subtracted 25ms. And also the parameter we set in
usleep() us 24920 by tuning, since usleep() cannot delay the program in
exact time parameter that you input.

Therefore, we decrease the original error in given 1.57
	joint position : 1.5769  ---> 1.5694
around 0.439% to 0.038% which increase 10% accuracy.