Fix selectVector bug in velocity based module
Because we use time synchruous method, if the joint we
select already reach the target velocity before we execute, the velocity
based module will never reach the final state and keep looping even if
all the joint reach the target velocity.
Therefore, we check the current velocityocity of each joint before run
the stop command and assign which joint is selected, so that RML velcoty
based algorithm could all deal with the selected joints.