Fix sequencial bug
When velocity based run complete, we will retrieve the current state
of robot from IP and feed to the next state run.
THerefore, in the end of velocity run and position run, we need to
re-point back the IP pointer : InputState = *IP, so that each run
function can return the current state when robot reach traget.